+++ docs/CHANGELOG.md
@@ -7,6 +7,8 @@ Released: ???
 Features:
 - Adapted module to the new fishing update:
   - Removed Guardian Pet alert and profit tracker as it's not obtainable from fishing anymore.
+  - Removed Shredder and Phantom Rod from Fishing Profit tracker.
+
 
 ## v1.30.0